// src/types/BreedingRecord.ts
export interface BreedingRecord {
  Parent1ID: string;
  Parent1Gender: string;
  Parent2ID: string;
  Parent2Gender: string;
  Parent3ID: string | null;
  Parent3Gender: string | null;
  Parent4ID: string | null;
  Parent4Gender: string | null;
  Parent5ID: string | null;
  Parent5Gender: string | null;
  Parent6ID: string | null;
  Parent6Gender: string | null;
  ChildID: string;
}